Move sub-commands to jam-server
- Fix some lints
